requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
asyncpg==0.29.0
attrs==23.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dataclasses-json==0.6.4
dnspython==2.6.1
fastapi==0.110.1
fastapi-sessions==0.3.2
fonttools==4.51.0
frozenlist==1.4.1
gigachain==0.1.16
gigachain-cli==0.0.21
gigachain-community==0.0.33.1
gigachain-core==0.1.43.1
gigachain-text-splitters==0.0.1
gigachat==0.1.23
gigaserve==0.0.51
gitdb==4.0.11
GitPython==3.1.43
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.25.2
httpx-sse==0.4.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.1.16
langchain-community==0.0.34
langchain-core==0.1.45
langchain-text-splitters==0.0.1
langsmith==0.1.49
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.4
mdurl==0.1.2
motor==3.4.0
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.1
packaging==23.2
pandas==2.2.2
pillow==10.3.0
plotly==5.21.0
psycopg2==2.9.9
pydantic==1.10.15
pydantic_core==2.18.1
Pygments==2.17.2
pymongo==4.6.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.29
sse-starlette==1.8.2
starlette==0.37.2
tenacity==8.2.3
tomlkit==0.12.4
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
uuid==1.30
uvicorn==0.23.2
yarl==1.9.4
